+++ tests/testthat/helper.R
@@ -21,7 +21,12 @@ expect_equal_or_both_error <- function(object, other, ...) {
       "tidypolars errored but tidyverse didn't."
     )
   } else {
-    if (is_polars_lf(polars_res)) {
+    if (isTRUE(other_error)) {
+      testthat::expect(
+        isTRUE(polars_error),
+        "tidyverse errored but tidypolars didn't."
+      )
+    } else if (is_polars_lf(polars_res)) {
       expect_equal_lazy(polars_res, other_res, ...)
     } else {
       expect_equal(polars_res, other_res, ...)